spurious compiler warning

+++ ATTINYCellModule/src/packet_processor.cpp
@@ -272,7 +272,7 @@ void PacketProcessor::onPacketReceived(PacketHeader *header)
     return;
   }
   // must be at least 1 because ::onHeaderReceived incremented it
-  if(badpackets == ~0)
+  if(badpackets == (uint16_t)~0)
     badpackets = 0;
   else
     badpackets += ((header->sequence - lastSequence) & 0x07) - 1;
